fix(android): proxy-only foreground start needs the specialUse FGS type
API 34 validates FGS types at startForeground time: systemExempted is only
permitted while the app is the active VPN (establish() ran), so proxy-only
mode crashed with ForegroundServiceStartNotAllowedException/
SecurityException on device. Declare specialUse alongside systemExempted
(with the required property + permission) and pass the mode-appropriate
type explicitly. Also pass gomobile's Go-int-as-long params correctly in
the logs/analytics calls (caught by the first on-device build), and record
the emulator smoke-test results in HANDOFF.

+++ android/app/src/main/java/com/webfilter/app/AnalyticsActivity.kt
@@ -58,7 +58,7 @@ class AnalyticsActivity : AppCompatActivity() {
     private fun refresh() {
         Thread {
             val result = try {
-                JSONObject(Mobile.analyticsJson(filesDir.absolutePath, hours))
+                JSONObject(Mobile.analyticsJson(filesDir.absolutePath, hours.toLong()))
             } catch (e: Exception) {
                 runOnUiThread {
                     Toast.makeText(this, getString(R.string.settings_save_failed, e.message ?: "error"), Toast.LENGTH_LONG).show()
